feat: preserve callable defaults
